requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
replicate>=0.25.0
google-genai>=0.2.0
cloudinary>=1.36.0
openai>=1.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
